feat: adopt material-aware SZLab workflows

+++ szlab_poly_studio/workflows/szlab_mixer_pump_production.py
@@ -9,7 +9,7 @@ szlab_mixer_pump: SzlabMixerPumpDevice = device("szlab_mixer_pump")
     workflow_uuid="127cf68e-43b3-58ab-932f-984f2b57019e",
     displayname="S06 加液生产流程",
 )
-def szlab_mixer_pump_production(*, pump: int = 1, volume: float = 8.0) -> None:
+def szlab_mixer_pump_production(*, pump: int = 1, volume: int = 8) -> None:
     # unilab:node_uuid=958e8abc-1cb6-5d75-b1cb-eaf0524b3a54
     addition = szlab_mixer_pump.run_solvent_addition(  # noqa: F841
         pump=pump,
